requirements.txt
aiogram==3.17.0
python-dotenv==0.19.2
yt-dlp==2025.2.19
